requirements.txt
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.26.0
boto3>=1.28.0
supabase>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
